Collapse Config, settings indirection and more (#42)
Various refactoring (as title says; collapsed two config classes), added settings.xml indirection, and lowered runtime requirement to Java 11 (IT runs only with 3.9.9 naturally).

Changes:
* added `dryRun`
* implemented `settings.xml` server/config indirection
* Lowered runtime requirement to Java 11
* collapsed `Config` and `SessionConfig`, `NjordSession` renamed to `Session`
* simplified session and config creation (now extension is mandatory to be installed to use Mojos)
* added Sisu-like ITs (mimicking what we want to achieve with Sisu)
* the `PublisherMojoSupport` heuristic improved
* added `deploy` publisher (does same thing as m-deploy-p does, but from locally staged repository)

This PR now allows setup like this:
0. set up Njord as described
1. add to `settings.xml` following (below is example for Sisu, that uses Sonatype Central Portal and has enabled snapshots)
```xml
    <server>
      <id>sisu-publish-service</id>
      <configuration>
        <njord.publisher>sonatype-cp</njord.publisher>
        <njord.releaseUrl>template:release-sca</njord.releaseUrl>
        <njord.snapshotUrl>template:snapshot-sca</njord.snapshotUrl>
      </configuration>
    </server>
```
2. edit POM to reflect "reality": `distributionManagement` should point to URLs from where published artifacts will be available once published (ie. central URL and Central Portal snapshots). Server ID of both should match the `sisu-publish-service` in `settings.xml`
3. perform `mvn deploy -P release` -- this will locally stage
4. perform `mvn njord:publish` -- this will publish

+++ core/src/main/java/eu/maveniverse/maven/njord/shared/impl/store/DefaultInternalArtifactStoreManager.java
@@ -10,7 +10,6 @@ package eu.maveniverse.maven.njord.shared.impl.store;
 import static java.util.Objects.requireNonNull;
 import static java.util.stream.Collectors.toList;
 
-import eu.maveniverse.maven.njord.shared.Config;
 import eu.maveniverse.maven.njord.shared.SessionConfig;
 import eu.maveniverse.maven.njord.shared.impl.CloseableConfigSupport;
 import eu.maveniverse.maven.njord.shared.impl.DirectoryLocker;
@@ -23,6 +22,7 @@ import eu.maveniverse.maven.njord.shared.store.RepositoryMode;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.net.URI;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 import java.nio.file.Files;
@@ -69,8 +69,8 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
     @Override
     public List<String> listArtifactStoreNames() throws IOException {
         checkClosed();
-        if (Files.isDirectory(config.config().basedir())) {
-            try (Stream<Path> stream = Files.list(config.config().basedir())) {
+        if (Files.isDirectory(config.basedir())) {
+            try (Stream<Path> stream = Files.list(config.basedir())) {
                 return stream.filter(Files::isDirectory)
                         .filter(p -> Files.isRegularFile(metaRepositoryProperties(p)))
                         .map(p -> p.getFileName().toString())
@@ -134,13 +134,15 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
         requireNonNull(name);
         checkClosed();
 
-        Path basedir = config.config().basedir().resolve(name);
+        Path basedir = config.basedir().resolve(name);
         if (Files.isDirectory(basedir)) {
             DirectoryLocker.INSTANCE.lockDirectory(basedir, true);
             try {
                 Path meta = metaRepositoryProperties(basedir);
                 if (Files.exists(meta)) {
-                    FileUtils.deleteRecursively(basedir);
+                    if (!config.dryRun()) {
+                        FileUtils.deleteRecursively(basedir);
+                    }
                     return true;
                 }
             } finally {
@@ -156,7 +158,7 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
         names.sort(Comparator.naturalOrder());
         Map<ArtifactStoreTemplate, TreeSet<String>> stores = new HashMap<>();
         for (String name : names) {
-            Path basedir = config.config().basedir().resolve(name);
+            Path basedir = config.basedir().resolve(name);
             Map<String, String> properties = loadStoreProperties(basedir);
             ArtifactStoreTemplate template = loadTemplateWithProperties(properties);
             stores.computeIfAbsent(template, k -> new TreeSet<>()).add(name);
@@ -166,7 +168,7 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
             for (String name : stores.get(template)) {
                 String formattedName = formatArtifactStoreName(template.prefix(), num++);
                 if (!formattedName.equals(name)) {
-                    Path basedir = config.config().basedir().resolve(name);
+                    Path basedir = config.basedir().resolve(name);
                     renameStore(basedir, formattedName);
                 }
             }
@@ -183,7 +185,7 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
             throw new IllegalArgumentException("Unsupported store type: " + artifactStore.getClass());
         }
 
-        Path targetDirectory = Config.getCanonicalPath(file);
+        Path targetDirectory = SessionConfig.getCanonicalPath(file);
         Path bundleFile = targetDirectory;
         if (Files.isDirectory(targetDirectory)) {
             bundleFile = targetDirectory.resolve(artifactStore.name() + ".ntb");
@@ -193,7 +195,8 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
         if (Files.exists(bundleFile)) {
             throw new IOException("Exporting to existing bundle ZIP not supported");
         }
-        try (FileSystem fs = FileSystems.newFileSystem(bundleFile, Map.of("create", "true"), null)) {
+        try (FileSystem fs =
+                FileSystems.newFileSystem(URI.create("jar:" + bundleFile.toUri()), Map.of("create", "true"), null)) {
             Path root = fs.getPath("/");
             if (!Files.isDirectory(root)) {
                 throw new IOException("Directory does not exist");
@@ -215,10 +218,11 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
         if (!Files.isRegularFile(file)) {
             throw new IllegalArgumentException("File does not exist");
         }
-        Path storeSource = Config.getCanonicalPath(file);
+        Path storeSource = SessionConfig.getCanonicalPath(file);
         String storeName;
         Path storeBasedir;
-        try (FileSystem fs = FileSystems.newFileSystem(storeSource, Map.of("create", "false"), null)) {
+        try (FileSystem fs =
+                FileSystems.newFileSystem(URI.create("jar:" + storeSource.toUri()), Map.of("create", "false"), null)) {
             Path repositoryProperties = metaRepositoryProperties(fs.getPath("/"));
             if (Files.exists(repositoryProperties)) {
                 Map<String, String> properties = loadStoreProperties(fs.getPath("/"));
@@ -254,7 +258,7 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
     }
 
     private DefaultArtifactStore loadExistingArtifactStore(String name) throws IOException {
-        Path basedir = config.config().basedir().resolve(name);
+        Path basedir = config.basedir().resolve(name);
         if (Files.isDirectory(basedir)) {
             DirectoryLocker.INSTANCE.lockDirectory(basedir, false);
             Map<String, String> properties = loadStoreProperties(basedir);
@@ -286,7 +290,7 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
 
     private DefaultArtifactStore createNewArtifactStore(ArtifactStoreTemplate template) throws IOException {
         String name = newArtifactStoreName(template.prefix());
-        Path basedir = config.config().basedir().resolve(name);
+        Path basedir = config.basedir().resolve(name);
         Files.createDirectories(basedir);
         DirectoryLocker.INSTANCE.lockDirectory(basedir, true);
         Instant created = Instant.now();
@@ -338,7 +342,7 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
     private String newArtifactStoreName(String prefix) throws IOException {
         String prefixDash = prefix + "-";
         int num = 0;
-        try (Stream<Path> candidates = Files.list(config.config().basedir())
+        try (Stream<Path> candidates = Files.list(config.basedir())
                 .filter(Files::isDirectory)
                 .filter(d -> d.getFileName().toString().startsWith(prefixDash))
                 .filter(d -> Files.isRegularFile(metaRepositoryProperties(d)))
@@ -370,6 +374,9 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
     }
 
     private void saveStoreProperties(Path basedir, Map<String, String> properties) throws IOException {
+        if (config.dryRun()) {
+            return;
+        }
         Properties prop = new Properties();
         properties.forEach(prop::setProperty);
         Path metaStoreProperties = metaRepositoryProperties(basedir);
@@ -380,6 +387,9 @@ public class DefaultInternalArtifactStoreManager extends CloseableConfigSupport<
     }
 
     private void renameStore(Path basedir, String newName) throws IOException {
+        if (config.dryRun()) {
+            return;
+        }
         Map<String, String> props = loadStoreProperties(basedir);
         props.put("name", newName);
         saveStoreProperties(basedir, props);
